CbiomesProcessing.jl documentation

CbiomesProcessing.jl documentation

CbiomesProcessing.jl provides a simple, distributed workflow to e.g. post-process binary model output. Inter-operability with MeshArrays.jl and NCTiles.jl is an important prospect of this package.

Contents

CbiomesProcessing.jl documentation

Contents

Main Features

Index

API / Functions

Move API here?

Main Features

Currently the reference implementation consists of

TaskDriver() (inside utilities.jl) is the driver function

It calls e.g. loop_task1.(i)

CbiomesProcessing.loop_task1(indx::Int) (inside fileloops.jl) loads M from MTRX.jld

It calls loop_task1(indx,M)

CbiomesProcessing.loop_task1(indx,M) (inside fileloops.jl) loops over files and variables

It calls MatrixInterp(in,MTRX,siz) and writes result to a FortranFile

API / Functions further documents the CbiomesProcessing module functions that are included in:

examples.jl includes simple examples without TaskDriver.

fileloops.jl includes task loops such as CbiomesProcessing.loop_task1()](@ref). A real-life example is provided in the TaskDriver documentation reported below.

utilities.jl provides low-level functions such as TaskDriver, MatrixInterp(), and MetaFileRead().

Index

CbiomesProcessing.MatrixInterp

CbiomesProcessing.MetaFileRead

CbiomesProcessing.StartWorkers

CbiomesProcessing.TaskDriver

CbiomesProcessing.cbioproc_example1

CbiomesProcessing.cbioproc_example2

CbiomesProcessing.loop_exampleA

CbiomesProcessing.loop_exampleB

CbiomesProcessing.loop_task1

CbiomesProcessing.loop_task1

CbiomesProcessing.loop_task2

CbiomesProcessing.loop_task3

CbiomesProcessing.loop_task4

CbiomesProcessing.prep_MTRX

CbiomesProcessing.read_SPM

API / Functions

CbiomesProcessing.MatrixInterp — Method.

MatrixInterp(in::Array{T,N},MTRX,siz) where {T,N}

Interpolate in using MTRX to grid of size siz.

source

CbiomesProcessing.MetaFileRead — Method.

MetaFileRead(filIn::String)

Reads a meta file generated by MITgcm

source

CbiomesProcessing.StartWorkers — Method.

StartWorkers(nwrkrs::Int)

Start workers if needed.

source

CbiomesProcessing.TaskDriver — Method.

TaskDriver(indx,fn)

Broacast / distribute task (fn; e.g. loop_task1) over indices (indx; e.g. file indices)

Examples:

using CbiomesProcessing, Distributed, SparseArrays
TaskDriver(1,CbiomesProcessing.loop_task1)

StartWorkers(4)
@everywhere using CbiomesProcessing, SparseArrays
TaskDriver(1:4,CbiomesProcessing.loop_task1)

Visualize results:

using FortranFiles, Plots
k=1
recl=720*360*4
fil="diags_interp/ETAN/ETAN.0000000732.data"
f =  FortranFile(fil,"r",access="direct",recl=recl,convert="big-endian")
tmp=read(f,rec=k,(Float32,(720,360))); close(f)
heatmap(tmp)

source

CbiomesProcessing.cbioproc_example1 — Method.

cbioproc_example1()

A 3D example without land mask or MeshArrays.

source

CbiomesProcessing.cbioproc_example2 — Method.

cbioproc_example2()

A 2D example with MeshArrays + mask or loop.

source

CbiomesProcessing.loop_exampleA — Method.

loop_exampleA(indx,SPM,siz)

Applies MatrixInterp in a loop over a subset of model output files (filList[indx])

source

CbiomesProcessing.loop_exampleB — Method.

loop_exampleB(indx,SPM,siz)

Applies MatrixInterp in a loop over a subset of model output files (filList[indx])

source

CbiomesProcessing.loop_task1 — Method.

loop_task1(indx,M)

Loop over a subset of model output files (filList[indx]), apply MatrixInterp (M) as a postprocessing step, and write the result to file (one subfolder for each variable)

source

CbiomesProcessing.loop_task1 — Method.

loop_task1(indx::Int)

Interpolate all variables for one record

source

CbiomesProcessing.loop_task2 — Method.

loop_task2(indx)

Loop over a subset of model output files (filList[indx]), convert Rirr* to Satellite Chl as a postprocessing step, and write the result to file (one subfolder for each variable)

source

CbiomesProcessing.loop_task3 — Method.

loop_task3(indx)

Loop over a subset of model output files (filList[indx]), compute export as a postprocessing step, and write the result to file (one subfolder for each variable)

source

CbiomesProcessing.loop_task4 — Method.

loop_task4(indx)

Loop over model output files and time average them as a postprocessing step, and write the result to file (one subfolder for each variable)

source

CbiomesProcessing.prep_MTRX — Method.

prep_MTRX()

Repackage interpolation matrix, mask, etc to .jld file.

source

CbiomesProcessing.read_SPM — Method.

read_SPM(dirIn::String)

Reads pre-computed interpolation (sparse matrix) from dirIn*"interp_precomputed.mat".

source